arreglo codigo
