+++ ParKingZone/src/main/java/pe/edu/upc/ParKingZone/serviceinterfaces/IMembresiaService.java
@@ -1,0 +1,15 @@
+package pe.edu.upc.parkingzone.serviceinterfaces;
+
+
+import pe.edu.upc.parkingzone.entities.Membresia;
+
+
+import java.util.List;
+
+public interface IMembresiaService {
+
+    public void insert(Membresia membresia);
+    public List<Membresia> list();
+
+
+}
