+++ ParKingZone/src/main/java/pe/edu/upc/ParKingZone/serviceinterfaces/IUsuarioService.java
@@ -1,0 +1,11 @@
+package pe.edu.upc.parkingzone.serviceinterfaces;
+
+import pe.edu.upc.parkingzone.entities.Usuario;
+
+import java.util.List;
+
+public interface IUsuarioService {
+
+    public void insert(Usuario usuario);
+    public List<Usuario> list();
+}
