+++ ParKingZone/src/main/java/pe/edu/upc/ParKingZone/servicesimplements/UsuarioServiceImplement.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pe.edu.upc.parkingzone.entities.Usuario;
 import pe.edu.upc.parkingzone.repositories.IUsuarioRepository;
-import pe.edu.upc.parkingzone.servicesinterfaces.IUsuarioService;
+import pe.edu.upc.parkingzone.serviceinterfaces.IUsuarioService;
 
 import java.util.List;
