--- ParKingZone/src/main/java/pe/edu/upc/ParKingZone/servicesinterfaces/IMembresiaService.java
@@ -1,15 +1,0 @@
-package pe.edu.upc.parkingzone.servicesinterfaces;
-
-
-import pe.edu.upc.parkingzone.entities.Membresia;
-
-
-import java.util.List;
-
-public interface IMembresiaService {
-
-    public void insert(Membresia membresia);
-    public List<Membresia> list();
-
-
-}
