--- ParKingZone/src/main/java/pe/edu/upc/ParKingZone/servicesinterfaces/IUsuarioService.java
@@ -1,11 +1,0 @@
-package pe.edu.upc.parkingzone.servicesinterfaces;
-
-import pe.edu.upc.parkingzone.entities.Usuario;
-
-import java.util.List;
-
-public interface IUsuarioService {
-
-    public void insert(Usuario usuario);
-    public List<Usuario> list();
-}
